Match LoginActivity_sec with Facebook login button && Sync facebook with firebase
